% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Scrape.R
\name{author_search}
\alias{author_search}
\title{Function to search google scholar by author and/or institution}
\usage{
author_search(author, institution = "", write = F)
}
\arguments{
\item{author}{Author's name. This works best in "lastname, firstname" format but any format is accepted (search results may vary)}

\item{institution}{Institution name, defaults to ""}

\item{write}{Write results to table? Can be TRUE or the name of the output file. Defaults to FALSE, which returns the results.}
}
\description{
Function to search google scholar by author and/or institution
}

