% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Scrape.R
\name{getBibtex}
\alias{getBibtex}
\title{Retrieve bibtex citation from google scholar, given the unique reference code}
\usage{
getBibtex(cite)
}
\arguments{
\item{cite}{Google scholar reference code}
}
\description{
Retrieve bibtex citation from google scholar, given the unique reference code
}

