% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Scrape.R
\name{scraper_internal}
\alias{scraper_internal}
\title{Assemble the data from a page of google search responses}
\usage{
scraper_internal(URL)
}
\arguments{
\item{URL}{link to google search responses}
}
\description{
Assemble the data from a page of google search responses
}

